Support option files and dynamically binding interfaces
This PR adds support for dynamically binding interfaces matching a regex. This
allows listening on interfaces in other L3 master domains, which the existing
bind strategy cannot handle.

Furthermore, this PR extends the "hostname override" concept to a file of
override options for other things like IPs and gateways. If a file named
`<interface>.options` exists in the same location as the existing override
files, it will be parsed and take precedence. See `options/options.go` for
details.