Support non /24 netmasks via the options file
This PR adds support for netmasks that are not /24. This is essentially
implemented by managing IPs as `net.IPNet` (address + mask) instead of plain
`net.IP`. Whenever a CIDR-less IP is encountered, a /24 is implicitly generated,
so the changes remain backwards compatible.

The PR also includes a couple of bug fixes, see inline PR comments.